move codedetailsloader to container-definitions
